Update Widget.cls_widget_names, check values
Some views, like patternfly tabs, have class attributes with immutable
values that are version picked.

In widget.cls_widget_names, only the attribute value was checked for
Widgetable, which VersionPick is.

This adds a 2nd check, that the items in value.child_items are
themselves Widgetable, to catch when a VersionPick object includes a
non-widgetable immutable.

Should not impact existing uses where value is Widgetable, or is a
VersionPick object that includes widgets.

This verifies ALL items in VersionPick are widgetable, and only includes
the widget in widget_names if so.