Fix for FF/Gecko driver bug
https://bugzilla.mozilla.org/show_bug.cgi?format=default&id=1411264
Apparently this will not be fixed and so we have to handle the fact
that the page changes and selenium returns early. In order to do this
we wait for the body to disappear and then reappear. This should happen
very quickly, but adding this onto every click is time consuming.
To mitigate this, clicks where we expect a page to need a change,
we pass in a new kwarg that explicitly targets this new code portion.